ai_agent/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
sentence-transformers
numpy
scikit-learn
python-dotenv==1.0.0
langchain-chroma==0.1.0
langchain-core==0.1.0
langchain-google-genai==0.0.6